backend/requirements/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2

psycopg[binary]==3.2.3
pgvector==0.3.6

celery==5.4.0
redis==5.2.1

gunicorn==23.0.0

# (later) LangChain/LangGraph will be added
# langchain
# langgraph
